backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langgraph==0.2.16
groq==0.4.1
websockets==12.0
pydantic>=2.7.4,<3.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
redis==5.0.1
asyncio-mqtt==0.16.1
python-dotenv==1.0.0